Add Tests for Virtual Thread Support for JVMTI Extensions API
In each of the tests, a virtual thread runs methodToCompile()
for a relatively large number of iterations to JIT compile the
method. Then, we use the JVMTI Extension API to check if there
are any JITted frames in the stack trace of the virtual thread.
We test this for both mounted and unmounted virtual threads.

Related: https://github.com/eclipse-openj9/openj9/issues/18386

Signed-off-by: Gengchen Tuo <gengchen.tuo@ibm.com>