Fix JITServer bug related to ChangesCurrentThread annotation
Methods that are annotated with @ChangesCurrentThread may receive
special treatment from the inliner. This is a feature present only
in Java21 and up. When creating a resolvedMethod mirror, the client
retrieves the information related to @ChangesCurrentThread annotation
and packs it into the message sent back to the server.
Current code contained a bug where the information related to this
annotation is not unpacked at the server. The net effect was that
the inliner refused to inline some methods, reducing throughput.
This commit rectifies the problem, by correctly unpacking the information
related to the ChangesCurrentThread annotation.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>