Treat invokespecial symbolic reference as superclass in method search
From JVMS: If the symbolic reference names a class (not an interface), then
that class is a superclass of the current class.

Previously the algorithm for finding the invoked method for invokespecial
assumes that the class named through the bytecode symbolic reference is the
current class when it is actually the superclass of the current class. By
doing this it skips searching direct interfaces of the referenced class for
default interface methods.

Signed-off-by: Theresa Mammarella <Theresa.T.Mammarella@ibm.com>