Implement more robust deserializer reset detection
Compilation threads must be able to detect whether or not a concurrent
JITServer AOT cache deserializer reset has occurred during the course of
deserializing a particular method. This check must be performed before
any lookup or store operation involving the deserializer's internal
caches. The methods of the deserializer now detect and report that
such a reset has occurred by first acquiring the relevant deserializer
monitor, checking the new _deserializerWasReset flag in the
ongoing compilation's TR::CompilationInfoPerThread, and setting a
&wasReset flag passed in to the method if such a reset occurred.

The _deserializerWasReset flag is set for every compilation thread
by the thread doing the resetting, when it is guaranteed to have
exclusive access to the deserializer's caches. Each compilation thread
will reset its own flag at the start of a remote compilation, when they
are guaranteed not to be processing JITServer AOT cache records from an
old server connection.

Signed-off-by: Christian Despres <despresc@ibm.com>