Modify SCC frontend to take method parameter
The new isMethodInSharedCache method in the shared class cache API
will determine whether or not the J9ROMMethod underlying a J9Method
is in the SCC, and will return its offset to the caller through an
input uintptr_t pointer if so.

The isROMMethodInSharedCache and offsetInSharedCacheFromROMMethod
methods are now forbidden from being called at the server. These should
only be called at a client, when working directly with the underlying
shared classes implementation.

Signed-off-by: Christian Despres <despresc@ibm.com>