Use server AOT cache instead of local SCC if requested
The methods of TR_J9JITServerSharedCache and the handling of the
well-known classes object in the SVM and AOT relocation header
initialization have been modified so that it we are compiling an
AOT cache store and are ignoring the client's local SCC, we now consult
the JITServer AOT cache and use the idAndType of its records in place
of local SCC offsets.

Various client-server messages have been modified to accommodate the
changes in server compilation. The client will either test for a local
SCC before accessing it, or the server will send an explicit boolean
parameter indicating that a local SCC offset is required.

Signed-off-by: Christian Despres <despresc@ibm.com>