Add var handle method type lookup table walking
The var handle method type lookup table is now walked in
allSlotsInROMClassDo(). This section of a J9ROMClass, written by
ROMClassWriter::writeVarHandleMethodTypeLookupTable() if
J9VM_OPT_METHOD_HANDLE is active, was previously not visited by the ROM
class walker. In addition to being necessary in general to walk this
section for the sake of completeness, walking this ROM class section is
necessary in particular so that JITServerHelpers::packROMClass() can
have access to all ROM class sections and their sizes. Not walking every
section of the ROM class led to a fatal assert being triggered in
packROMClass in the related issue.

Fixes: #18957

Signed-off-by: Christian Despres <despresc@ibm.com>