Recompile the outermost method in JIT dump when compilation crashes
Previously the JIT dump would recompile the "current method," which is
sensitive to ongoing IL generation. If the compiler crashed during IL
generation for an inlined method, then only that inlined method would be
recompiled.

Now the JIT dump will always recompile the outermost method so that
there will be a chance of reproducing IL generator crashes even if they
are related to inlining.