Do not throw in ClassLoaderTable
In copyJ9UTF8(), we now avoid throwing std::bad_alloc() if memory for a
copy of a ROM class's name could not be allocated. We instead do not
add the TR_ClassLoaderInfo into the _nameTable in such situations. This
parallels the treatment of the class loader info if a chain for the
loader could not be found in the SCC.

Signed-off-by: Christian Despres <despresc@ibm.com>