Cache compInfoPT in JITServer AOT deserializer SCC
Caching the per-thread compilation thread info saves looking it up every
time the TR_J9DeserializerSharedCache needs to obtain the current
compilation object.

To support this caching, there is now a TR_J9DeserializerSharedCache
in TR_J9VMBase, which is initialized when a frontend is created for a
particular compilation thread.

Signed-off-by: Christian Despres <despresc@ibm.com>