Enable per-connection JITServer AOT cache disabling
If a JITServer client uses -XX:+JITServerAOTCacheIgnoreLocalSCC,
servers that client connects to will now

1. Fail a compilation completely if the client requests an AOT cache
   store and the server cannot set up the compilation as an AOT cache
   store or load, and

2. Report the state of the server's AOT cache if (1) occurs, so the
   client will not continue to send AOT cache requests to that server
   if those requests could not possibly be fulfilled.

The client tracks the state of the current server's AOT cache with
the persistent info properties doNotRequestJITServerAOTCacheStore() and
doNotRequestJITServerAOTCacheLoad(). Since a particular server's cache
will never become able to complete AOT cache loads or stores partway
through a connection (e.g., the server's cache never shrinks in size,
so AOT cache stores will never succeed once it is full) these properties
remain true once set for the duration of the connection with a
particular server. They are reset whenever a client connects to a new
server.

Signed-off-by: Christian Despres <despresc@ibm.com>