Refuse to generate array class serialization records
The JITServer AOT cache will now refuse to generate class records for
array ROM classes. This is relevant for clients running with
-XX:+JITServerAOTCacheIgnoreLocalSCC - with this option disabled, the
fact that array classes cannot be persisted in the local SCC acts as
an implicit barrier to their ROM classes ever reaching the AOT cache.
Without that option, the AOT cache must explicitly check that the given
ROM class is not an array ROM class.

In future, the AOT cache should be extended with support for array
classes. For now, this change prevents excessive AOT cache
deserialization failures from occurring at JITServer clients running
with -XX:+JITServerAOTCacheIgnoreLocalSCC due to the nonsensical
class records the cache previously generated for array classes.

Signed-off-by: Christian Despres <despresc@ibm.com>