Update ScopedMemoryAccess_closeScope0 for JDK22
In JDK22, the definition has changed to

native void
closeScope0(MemorySessionImpl session, ScopedAccessError error);

The return type has changed from boolean to void, and a second
parameter is re-introduced: "ScopedAccessError error".

Previously, closeScope0 would be retried until it returned true,
which suggested that the scope/session was not found on any thread's
stack.

The new behaviour sets an asynchronous exception for all threads,
which have the scope/session on their stack. The second parameter,
error, is the asynchronous exception, which is thrown. Before
throwing the error, the thread verifies if the scope/session is
still on the stack. If not, the asynchronous exception is cleared
and not thrown.

Related: https://github.com/eclipse-openj9/openj9/issues/13211#issuecomment-1984768250

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>