Ensure visibility of post-<init> field values for StringBuffer/Builder
When running on weakly ordered machines like PPC or AArch64 it is possible for non-final fields to be seen as NULL by other thread even when the fields are initialized by the constructor (<init>). In such cases it's up to the java program to use some sort of ordering mechanism if the results of the <init> must be seen by other threads before those threads are allowed to see the new object.

For StringBuffer and StringBuilder the JIT is removing null-checks to improve performance and therefore it is possible (for programs that fail to properly enforce ordering when multiple threads can see a new StringBuffer/StringBuilder instances) to crash trying to write to NULL.

One possible solution is to stop removing null-checks for those object types, but that would introduce a performance loss in order to avoid the crash. It would also result in a rare NullPointerException on strongly ordered machines running poorly written java code. This solution is a perfectly valid one that is consistent with the JVM specification, but does not improve the behaviour consistency between CPU architectures.

The solution this fix implements is to move the allocation fence from post allocation (pre <init>) to post <init> for both StringBuffer and StringBuilder object instances. This is a safe change only because we know those two classes no not allow the 'this' pointer to escape to other threads when executing the <init>. Therefore, the the pre <init> fence is not required knowing that the new object can't been seen by other threads until after executing the <init>, at which point we will be executing the fence. This makes it impossible for a pre <init> field state to be seen by other threads, and by extension making it impossible for a NULL to be seen. In this way the JIT can continue to remove null-checks without the possibility of encountering a crash due to NULL pointers, and at the same time avoid NullPointerException's on weakly ordered machines making a small improvement to the consistency between CPU architectures for the JVM. And since we are only moving the fence (not adding one) we introduce no performance penalty.