Allow JITServer communication and health port 0
The -XX:JITServerPort=<port> and -XX:JITServerHeathProbePort=<port>
options both now support the port value 0. When this port is specified,
a random unassigned port will be assigned to the relevant socket, and
this bound port will be reported in the JITServer verbose log.

This behaviour when port 0 is specified matches that of the underlying
bind() network function.

Signed-off-by: Christian Despres <despresc@ibm.com>