Fallback to UTF8 converting command line arguments
On Linux the encoding of the terminal seems to be unknown and isn't
necessarily `locale charmap`. When converting command line arguments and
the conversion fails using the default encoding, try UTF8 as well.

OpenJDK accepts UTF8 input.

A way to test this is to have a locale such as `zh_CN.gb18030`
installed, and then run `LANG=zh_CN.gb18030 java -Dzz=〢€郎嗀秊㐀龦
-XshowSettings:properties -version`. The zz= characters should be UTF8
encoded. In a terminal using UTF8, and without this change the JVM fails
to start with `Error: Could not create the Java Virtual Machine.`.

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>