DDR: always generate optional fields
The test of whether a field is present must be deferred until a
specific system dump is being accessed.

Structure stubs must include all fields optional or otherwise. The
structure class generated dynamically will only include those fields
actually present in the JVM that produced the system dump. Accesses
to missing fields will trigger a NoSuchFieldError which will be
replaced by NoSuchFieldException by the related pointer class.

Pointer classes must support the scenario where a field is not present
in the executing JVM, but is present in a system dump being analyzed.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>