Refine JITServer AOT cache flag setting
When using the JITServer AOT cache and ignoring the local SCC, portable
AOT is now enabled by default only if the JVM is running in a container,
and only if there is no explicit -XX:[+|-]PortableSharedCache option
present on the command line.

The -XX:[+|-]PortableSharedCache options are also now honoured when
using the JITServer AOT cache (having the effect of explicitly enabling
or disabling portable AOT) even if the SCC is disabled.

Signed-off-by: Christian Despres <despresc@ibm.com>