Relax condition on JITServer rememberClass assert
When ignoring the client's SCC, if a server performs an AOT compilation
that is not an AOT cache compilation for a client (possible if the
client has a valid local SCC) it may cache valid class chain offsets for
classes in the client's session data. If the server then performs an
AOT cache store for that client, it may receive an invalid class chain
from the client for a class that already has a valid cached class chain
offset, since the client will be ignoring its local SCC for such a
compilation. This situation is valid, and should not trigger the fatal
assert in TR_J9JITServerSharedCache::rememberClass() that detects a
mismatch between the cached and received class chain offset. That fatal
assert is now disabled for AOT cache stores when the server is ignoring
the client's SCC.

This fixes an unnecessary crash that could occur in the situation
outlined above.

Signed-off-by: Christian Despres <despresc@ibm.com>