Optimize OpenJDK VarHandle operation methods
VarHandle operation methods are mostly calls to Unsafe methods,
which under certain circumstances can be folded into simple loads
and stores in UnsafeFastPath as we can determine the object base
and offsets at compile time. In other cases, we can unlock the
fast paths for the Unsafe operations by setting the appropriate
flags for CodeGen to inline the intrinsics if the platform supports
the operations.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>