Fix acquireVMaccessIfNeeded for Non-Comp Threads
Whena a non-comp thread invokes acquireVMaccessIfNeeded, the function
silently returns, because it assumes that a non-comp thread should
already have had VMAccess. This is particularly problematic when using
TR::VMAccessCriticalSection, as it is not at all obvious that the
current thread may not actually have VMAccess. Furthermore, there are
legitmate circusmtances when a non-comp thread will not have VMAccess
and will need to use this API to acquire it.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>