[Verifier] Prevent displaying the garbage type on the empty stack
The changes aim to rectify the situation when there is
literally no valid type data to be displayed on the empty
stack after a bunch of manipulations (e.g. pop) to prevent
any garbage data from being decoded in the error message
framework so as to match the RI's behavior.

Signed-off-by: ChengJin01 <jincheng@ca.ibm.com>