Don't try compiling again during Shutdown
On shutdown, a compilation is aborted via the
TR::CompilationInterrupted exception. However, the error code associated
with this exception (compilationInterrupted) will result in the
compilation being retried.

In a synchronous compilation, it is possible for a thread to remaining
waiting on the queue slot monitor; it would only be notified if the
compilation was not retried. Essentially the following is possible in a
synchronous compilation:

* Requesting Thread adds method to be compiled
* Requesting thread waits on the queue slot monitor
* Compilation Thread picks up the entry, and starts compiling
* Shutdown Thread starts shutdown process which:
  * Interrupts compilations in progress
  * Sets the _compilationThreadState to COMPTHREAD_SIGNAL_TERMINATE
* Compilation Thread aborts compilation via TR::CompilationInterrupted
* Compilation Thread calls shouldRetryCompilation which:
  * Returns true for compilationInterrupted
* Compilation Thread requeues entry, does not notify waiting Requesting
  Thread
* Compilation Thread exits loop in
  TR::CompilationInfoPerThread::processEntries
* Requesting Thread remains waiting on the queue slot monitor

This commit updates shouldRetryCompilation to return false if the JVM is
shutting down.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>