Memory disclaim support for JIT persistent memory
This commit adds memory disclaim support for segments of
JIT persistent memory.
The `PersistentAllocatorKit` adds one more field which denotes
memoryType attributes. If the caller uses `MEMORY_TYPE_VIRTUAL`
for the PersistentAllocatorKit used in the construction of the
PersistentAllocator object, then the PersistentAllocator will
use mmap instead of malloc, and it will enable disclaiming.
If disclaiming to a backing file is desired (as opposed to
disclaiming to swap) then the caller also needs to pass the
`MEMORY_TYPE_DISCLAIMABLE_TO_FILE` flag to the PersistentAllocatorKit.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>