Improve code cache allocation time
The JIT attempts to allocate its code cache repository in close
proximity of the JIT dll to avoid helper trampolines.
A. If large pages are not used, the JIT requests an allocation in specified
range with `J9PORT_VMEM_ADDRESS_HINT`. This performs a single mmap call and
typically succeeds. If the call above did not reserve memory in the desired
range, we unmap and then instruct the VM to use a "quick" method based on
smaps. If that fails too, we ask for an allocation anywhere.
B. If large pages are used, the JIT requests an allocation in the desired
range and may repeat this allocation request after advancing the start
address by one page each time it fails. If we still didn't allocate a
segment in the desired region, we ask for an allocation anywhere.

The problem fixed by this commit was that the "large page" path was
always taken. The code was comparing the value of `largeCodePageSize`
against 0, but `largeCodePageSize` is set to 4 KB when large pages are not
used.
This commit changes the code to correctly set `largeCodePageSize` to 0
when the default page size is used for code caches.

Signed-off-by: Marius <mpirvu@ca.ibm.com>