Rework processing of Java system properties in java.lang.ClassLoader
The following Java system properties are impacted:
- reflect.cache
- java.lang.ClassLoader.lazyInitialization

toLowerCase() generates a LambdaForm for the associated invokedynamic.
LambdaForm generation code relies on ClassLoader.bootstrapClassLoader,
which is NULL and no longer initialized in JDK23. This causes a NPE.

In JDK23+, the above two Java system properties are no longer processed
with toLowerCase() since they are undocumented and only used for
debugging. This prevents a NPE.

Related: #19301

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>