Fix JITServer AOT cache compilation retry logic
When checking for the eligibility of a method to be JITServer AOT cache
stored or loaded, we now check that the flag _doNotAOTCompile
is not set in all cases. (This flag was also renamed, from the somewhat
misleading _doNotUseAotCodeFromSharedCache). This has the effect of
ensuring that when any unrecoverable AOT-related method compilation
failures occur, JITServer clients will not request that those methods be
JITServer AOT cache compiled during a subsequent compilation.

Signed-off-by: Christian Despres <despresc@ibm.com>