Recognize unsafe unaligned getters/setters
The commits adds support for unaligned getter/setters in the
Unsafe class for the JIT.

Unaligned getters and setters are treated the same as their
aligned equivalents for ISAs which permit unaligned reads/stores.
This requires a change to the inliner to consider whether
to emit an unsafe store/read based on whether the target ISA permits
unaligned accesses. So far only the Z target does not permit unaligned
accesses, so the getters/setters will fall back to their Java
software implementation.

Signed-off-by: James You <james.you@protonmail.com>