JITServer AOT cache array class support
This commit enables AOT cache serialization and load for methods that
refer to array classes in their relocation and validation records,
specifically object arrays and multidimensional primitive arrays. Since
such classes share the same ROMClass named "[L", we disambiguate them
in the AOT cache by using the full array class signature instead of the
ROMClass name. We build the ROMClass hash for an array class by
combining the ROMClass hash of the "[L" object array class, the ROMClass
hash of the base component class, and the number of array dimensions.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>