Fix fatal assertion in class loader table during class unload
When removing entries from the class loader table during class unloading, the
current implementation uses the wrong shared cache API isPointerInSharedCache()
(which is only intended for the metadata section of the SCC) to check whether
the class loader identifying name string is stored in the SCC or in the
persistent allocator heap. This can result in a fatal assertion when using both
the JITServer AOT cache and the local SCC, as well as memory corruption since
SCC memory ranges get added to the persistent allocator free lists. This commit
fixes the issue by using the correct API isPtrToROMClassesSectionInSharedCache().

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>