Fix resetting global lists issue in CRIU reinitializeForRestore
we need to reset the root of global lists(unfinalizedObjectLists,
OwnableSynchronizerObjectLists, ContinuationObjectLists) for rebuilding
the lists(array of sublists, the size of array match to gcThreadCount),
in case the gcThreadCount in restore env is more than the gcThreadCount
in original env.
Backup and reset root of global lists before reinitializeForRestore
and restore the root of global lists if the lists were not rebuilt
during reinitializeForRestore.

Signed-off-by: hulin <linhu@ca.ibm.com>