Stop predicting VH.asDirect()/directVarHandleTarget() during inlining
The MethodHandle can be found without knowing the result of this call
because it's looked up based on the original VarHandle reference, not
the one from directVarHandleTarget(), and this is enough to allow the
inliner to inline the entire access.

Additionally, the compiler has been failing to determine the result
anyway. The logic deleted here would give up whenever the VarHandle was
an instance of a subclass, which it seems is always the case.

Nothing looks for directVarHandleTarget() anymore, so it will no longer
be recognized. But asDirect() will continue to be recognized for
isJSR292SmallGetterMethod().