Remove unused method findOrCreateClassAndDepthFlagsSymbolRef
A method named findOrCreateClassDepthAndFlagsSymbolRef was introduced
by a previous commit to correct the typo in the name of the method
findOrCreateClassAndDepthFlagsSymbolRef.  That change kept the old
method due to upstream uses in J9_PROJECT_SPECIFIC code in OMR.  This
removes the old method name now that the upstream uses have been
removed.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>