Removes incorrect check in fast path indexOf code
For the recognized method JITHelpers.intrinsicIndexOfLatin1, the ch parameter
is meant to be interpreted as an unsigned byte with a value of 0x00-0xFF.
However, it gets passes into the method as a signed byte. This means for
values 0x80-0xFF the data is sign extended inside of the register.

The fast path code generated by inlineIntrinsicIndexOf_P10 has a check the
checks if the bits other than the lowest 8 bits are 0 and returns -1 early if
this is the case. But, this check is wrong. It is possible to have a valid
parameter where there upper bits are not 0 due to the sign extension.

This fix removes the check and instead zeros out the bits other than the
lowest 8 bits. This is safe to do since it is known in advance that the byte
data is meant to be interpreted as an unsigned byte and it is not possible to
pass in an invalid value to this parameter. Also, this is needed since the
byte parameter is eventually compared against a loaded byte that was zero
extended.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>