Support lambdas and other generated classes in JITServer AOT cache
This commit implements support for correctly and reliably
serializing and deserializing/loading JITServer AOT cache methods
that are defined by or refer to classes generated at runtime.

The current list of recognized types of generated classes includes:
- Lambda classes;
- Dynamic proxy classes;
- Generated reflection accessors.
In each of these cases, the generated class name consists of a
deterministic prefix followed by a suffix that can vary across JVMs.

Generated classes are identified across client JVMs using the
combination of: the class loader ID (name of its first loaded
class), the deterministic class name prefix, and the deterministic
ROMClass hash. The latter is computed after re-packing the
ROMClass to truncate all the instances of the class name string
down to the deterministic class name prefix. In order to lookup
generated classes at the client JVM during AOT deserialization
and load, we maintain a mapping between <loader, prefix, hash>
and the RAMClass, which is populated in the class load JIT hook,
and invalidated in the class unload and redefinition hooks.

This mechanism can be disabled by passing the command
line option `-Xjit:aotCacheDisableGeneratedClassSupport`
to both the JITServer and the client JVMs.

Signed-off-by: Alexey Khrabrov <khrabrov@cs.toronto.edu>