Add optimization to ROM class search for Anonymous classes unloading
Currently there is a N-square search behaviour during Anonymous classes
unloading - search for the segment where ROM class for unloaded RAM
class is stored. This general segments walk can introduce delays if
there is a large number of Anonumous classes loaded.
This PR introduces fast path can be attempted to find ROM class. There
is a high possibility that the ROM segment we are looking for is the
next one for RAM class segment. So, this segment can be checked first.
If fast path search was not succesful an existed slow path is going to
be used. Also can be a case when ROM class is not stored in Anonymous
classloader, for example for Indexable, there is no reason to even
search it.

Signed-off-by: Dmitri Pivkine <Dmitri_Pivkine@ca.ibm.com>