Expand Java 17+ aggressive SFFF from VarHandle to all types
Starting in Java 17, it is no longer possible to remove the final
attribute of a field to make it writable via reflection, and in classes
with version 53 and later, putstatic can write to static final fields
only during class initialization.

With this justification, the JIT compiler has been treating static final
fields with declared type VarHandle as reliable for folding (if declared
in a suitable class).

But the justification applies regardless of the declared type of the
static final field, so with this commit, so too will the folding.