Move anchored instanceof to the correct position
This commit includes the following two changes in ILGen:

1.
If the class is unresolved, `genInstanceof` anchors `instanceof`
under a `treetop` node. Then the anchored `instanceof` treetop
appears after the callTree and it commons with
the `instanceof` node under the `ZEROCHK` in `geninterface`.

This causes a problem because `expandUnresolvedClassInstanceof`
does not expect the treetop that has the anchored `instanceof`
to have already been evaluated when it appears under ZEROCHK.
The transformation in `expandUnresolvedClassInstanceof` will
not function correctly.

Therefore, the anchored `instanceof` treetop needs to be moved
up before ZEROCHK.

2.
If the receiver of the `instanceof` is non-null, there is no need
to transform the `instanceof` into two cases: the null case and
the non-null cases in `expandUnresolvedClassInstanceof`.

Fixes: #18748

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>