Enabled AOT cache at server by default
This change makes the option -XX:+JITServerUseAOTCache unnecessary at
the server; note that clients are still required to set that option to
opt-in to the use of the AOT cache. The -XX:-JITServerUseAOTCache option
may still be used at the server to disable the AOT cache as before.

This change also has the effect of making ROM class sharing at the
server the default.

Signed-off-by: Christian Despres <despresc@ibm.com>