Skip BigInteger zero-init & checks for OpenJ9 only
The BigInteger.toByteArray implementation in OpenJ9 doesn't depend
on zero-initialization and it is currently safe to skip that operation
in that context. The Apache Harmony BigInteger implementation relies
on zero-init and returns the wrong result without it. This patch limits
skipping of zero-init (and NULL and bounds checks, to be conservative)
for BigInteger.toByteArray to OpenJ9 builds.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>