Add dontFoldStaticFinalFields={} JIT option
This can be used to prevent constant folding a specified set of static
final fields for debugging purposes or as a workaround.

Whenever the JIT attempts to fold a static final field, if this option
is set, it checks whether the glob pattern ("simple regex") matches
<class>.<field>:<sig>, e.g.

    java/util/AraryList.EMPTY_ELEMENTDATA:[Ljava/lang/Object;

If it matches, then the static final field is not folded.